회원탈퇴할 때 비밀번호 입력하게 함

+++ src/app/profile/server.ts
@@ -6,18 +6,19 @@ export async function getProfile(username: string) {
   return user;
 }
 
-export async function deleteUser(username: string) {
-  await User.deleteByUsername(username);
+export async function deleteUser(username: string, password: string) {
+  const user = await User.findByUsername(username);
+  if (user?.password == password) {
+    await User.deleteByUsername(username);
+    return true;
+  }
+  return false;
 }
 
 export async function editProfile(username, name, birth, message) {
   await User.editProfile({ username, name, birth, message });
 }
 
-export async function editPassword(username,password) {
-  await User.editPassword({ username, password});
+export async function editPassword(username, password) {
+  await User.editPassword({ username, password });
 }
-
-
-
-
